# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Vol d'Histoire API
  description: |
    API REST de l'encyclopédie de l'aviation militaire **Vol d'Histoire**.

    ## Authentification
    JWT en header `Authorization: Bearer <access_token>` (15 min) +
    refresh token en cookie HttpOnly (7 jours, rotation au refresh).

    ## Internationalisation
    Détectée via `?lang=fr|en` ou `Accept-Language`. Les fiches d'avions
    exposent `name_fr` / `country_name_fr` / `type_name_fr` en plus des
    champs traduits, comme **clés stables côté client** (utilisées dans les
    URLs de filtres pour rester valides indépendamment de la langue d'affichage).

    ## Anti-spam
    Les routes `/register` et `/auth/forgot-password` acceptent un token
    hCaptcha dans le body (`h-captcha-response`). Si `HCAPTCHA_SECRET`
    n'est pas configuré côté serveur, la vérification est skippée (mode dev).

    ## Cache & versioning
    Toutes les pages HTML sont servies avec un query string `?v=<version>`
    sur les assets pour bust le cache navigateur à chaque déploiement.
    Les statiques (CSS/JS/fonts) ont un `Cache-Control` long.
  version: 4.0.0
  contact:
    name: Vol d'Histoire
    url: https://vol-histoire.titouan-borde.com
servers:
  - url: /api
    description: API principale
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token JWT (15 min). Obtenu via POST /login ou POST /refresh.
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
          format: email
        role_id:
          type: integer
          enum: [1, 2, 3]
          description: "1=admin, 2=éditeur, 3=membre"
...
  parameters:
...
    UserId:
      name: id
      in: path
      required: true
      schema:
        type: integer
paths:
...
  /users:
    get:
      tags: [Users]
      summary: Lister les utilisateurs (admin)
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: Liste paginée
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
                  meta:
                    type: object
                    properties:
                      total:
                        type: integer
                      page:
                        type: integer
                      limit:
                        type: integer
                      totalPages:
                        type: integer
        '403':
          description: Accès réservé aux admins
  /users/{id}:
    get:
      tags: [Users]
      summary: Profil utilisateur
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: Profil
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '403':
          description: Accès non autorisé (sauf son propre profil ou admin)
        '404':
          description: Utilisateur non trouvé

    put:
      tags: [Users]
      summary: Modifier son profil
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
      responses:
        '200':
          description: Mise à jour réussie
        '400':
          description: Données invalides
        '403':
          description: Accès non autorisé

    delete:
      tags: [Users]
      summary: Supprimer un compte
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: Compte supprimé
        '403':
          description: Accès non autorisé
  /admin/users/{id}:
    put:
      tags: [Users]
      summary: Modifier un utilisateur (admin)
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, role_id]
              properties:
                name:
                  type: string
                email:
                  type: string
                  format: email
                role_id:
                  type: integer
                  enum: [1, 2, 3]
      responses:
        '200':
          description: Utilisateur modifié
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '403':
          description: Accès réservé aux admins
        '409':
          description: Email déjà utilisé

  # ═══════════════════════════════════════════════════════════
  # AIRPLANES
  # ═══════════════════════════════════════════════════════════
...